# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API — Sistema de Pasantías IT Beltrán",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Desarrollo local"
    },
    {
      "url": "{publicUrl}/api",
      "description": "Despliegue",
      "variables": {
        "publicUrl": {
          "default": "https://api.ejemplo.edu"
        }
      }
    }
  ],
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT de sesión en cookie **HttpOnly `token`** (SEC-02). Se setea al hacer `POST /auth/login` y el navegador la reenvía sola. El JS del cliente no la ve."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Alternativa para clientes de API y esta consola: `Authorization: Bearer <token>`. El body de `POST /auth/login` devuelve el `token` a usar acá."
      }
    },
    "parameters": {
      "pageParam": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "Número de página (base 1).",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "limitParam": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Ítems por página. Se recorta al máximo del endpoint (12–100 según el caso).",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 20
        }
      },
...
    },
    "responses": {
      "ValidationError400": {
        "description": "Body inválido — falta un campo requerido o un valor no es válido.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Faltan campos obligatorios."
            }
          }
        }
      },
      "Error401": {
        "description": "No autenticado — falta el token, es inválido o expiró.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Token inválido o expirado."
            }
          }
        }
      },
      "Error403": {
        "description": "Sin permisos para la acción (rol de sistema o rol interno de empresa).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "No tenés permisos para realizar esta acción."
            }
          }
        }
      },
...
      "Error404": {
        "description": "El recurso no existe o no es visible para este usuario.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Recurso no encontrado."
            }
          }
        }
      },
...
      "Error429": {
        "description": "Límite de tasa alcanzado (rate limiting).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "code": "RATE_LIMITED",
              "message": "Demasiadas solicitudes. Esperá un momento e intentá de nuevo."
            }
          }
        }
      },
      "Error500": {
        "description": "Error interno no controlado.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Error interno del servidor.",
              "requestId": "a1b2c3d4-..."
            }
          }
        }
      }
    },
    "schemas": {
      "Envelope": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "const": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "message": {
            "type": "string",
            "description": "Mensaje legible en español."
          },
          "code": {
            "type": "string",
            "description": "Código tipificado (no siempre presente). Middleware: `CSRF`, `RATE_LIMITED`, `SIN_EMPRESA`, `SIN_ROL`, `ROL_INSUFICIENTE`. Negocio: `CV_REQUERIDO`, `POSTULACION_DUPLICADA`, `OFERTA_NO_ACTIVA`, `OFERTA_VENCIDA`, `EMAIL_DUPLICADO`, `YA_ES_MIEMBRO`, `EMAIL_REGISTRADO`, `EMAIL_SOLICITUD_PENDIENTE`, `PERFIL_PRIVADO`."
          },
          "rolesPermitidos": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Solo en 403 `ROL_INSUFICIENTE`: roles internos que sí pueden."
          },
          "requestId": {
            "type": "string",
            "description": "Solo en 500 — correlaciona con los logs técnicos."
          }
        }
      },
      "Pagination": {
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "description": "Página actual (recortada a [1, totalPages])."
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer",
            "description": "Total de filas del scope filtrado."
          },
          "totalPages": {
            "type": "integer",
            "description": "Siempre ≥ 1."
          }
        }
      },
...
      "UsuarioResumen": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "rol": {
            "type": "string",
            "enum": [
              "alumno",
              "egresado",
              "empresa",
              "admin"
            ]
          },
          "telefono": {
            "type": [
              "string",
              "null"
            ]
          },
          "ubicacion": {
            "type": [
              "string",
              "null"
            ]
          },
          "fotoPerfil": {
            "type": [
              "string",
              "null"
            ]
          },
          "ultimoAcceso": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "Mensaje": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "emisorId": {
            "type": "integer"
          },
          "receptorId": {
            "type": "integer"
          },
          "mensaje": {
            "type": "string"
          },
          "leido": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/chat/{usuarioId}": {
      "get": {
        "tags": [
          "chat"
        ],
        "operationId": "chatHistorial",
        "summary": "Historial de mensajes con un usuario",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Envelope"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data",
                        "pagination"
                      ],
                      "properties": {
                        "usuario": {
                          "$ref": "#/components/schemas/UsuarioResumen"
                        },
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Mensaje"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "parameters": [
          {
            "name": "usuarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          }
        ]
      }
    },
...
  }
}